backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
redis==5.0.1
influxdb-client==1.39.0

# Trading APIs
alpaca-py==0.17.0
ccxt==4.2.25
python-binance==1.0.19

# Data & Analysis
pandas==2.1.4
numpy==1.26.3
ta==0.11.0
yfinance==0.2.35

# AI/ML
torch==2.1.2
scikit-learn==1.4.0
openai==1.10.0
transformers==4.37.0

# News & Sentiment
feedparser==6.0.10
beautifulsoup4==4.12.3
requests==2.31.0
textblob==0.18.0
newspaper3k==0.2.8

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==42.0.0

# Utilities
python-dotenv==1.0.0
aiohttp==3.9.1
websockets>=11.0.3,<12.0.0
schedule==1.2.1
pytz==2023.3
python-dateutil==2.8.2

# Monitoring & Logging
prometheus-client==0.19.0
sentry-sdk==1.39.2

# Notifications
python-telegram-bot==20.7
twilio==8.11.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

